+++ lib/ui/worker/call.dart
@@ -743,6 +743,7 @@ class CallWorker extends Dependency {
       _outgoingAudio = AudioUtils.play(
         AudioSource.asset('audio/$asset'),
         fade: fade ? 1.seconds : Duration.zero,
+        mode: AudioMode.call,
       );
       previous?.cancel();
     } else if (asset == _endCall) {